refactor: rewrite logic in `src/main.rs`
I have done a big work with Waller. With the new
collection system and new design choices, I rewrite
`main.rs` to make it work better.

I have an idea to remove indexes and leave only
the names of the wallpapers in collection.
Becaus working with indexes and names at the same
time is so complicated and can lead to make code
look more like spagetti.